Implemented parts of MovementSystem.
MovementSystem should now be able to maintain a list of all theoretically possible moves all of the pieces on the board can make with the current configuration.

The classes that implemented the previously called MovementComponent interface were renamed, and a new MovementComponent component was introduced that acts as an aggregator for MovementRules.

A few accessory classes were created to support the MovementSystem like SizeComponent to get the size of the board, and Position class to represent a 2d point on the board.