Added support for Castling for king pieces.
Actual castling mechanisms will have to be implemented in-engine, so not sure how that will work yet, but I'm considering using a conditional move flagging system (i.e. when a piece does a conditional move it will notify the game somehow, and the game can use that to support more rich turns). This system will likely also be used for promotion of pawns, hence why I'm removing that TODO as well.