Updated Piece and Turn.
Turn is now closer to the planned turn syntax, which also makes it easier to work with.

Piece was cleaned up and most relevant functionality should now be in place with good comments, etc.

The PieceFactory was reworked into MovementFactory, and scope was reduced (Pieces are a bit too complex to be created without expanding the MovementFactory, so I chose to limit scope instead, at least for now.)