Fixed a couple of levelling bug.
The Piece class was incorrectly comparing it's XP against a threshold (it was checking if the XP it received was higher than the threshold rather than checking it's own XP against the threshold).

The LevelUpEffectFactory was incorrectly setting level-up effects for the bishop and knight, so that was fixed as well.

Not bug fixes:
Also added resetting level for the Queen when it "levels up".

Added toString to MovementRuleSet used when I was debugging.